\name{LPG-package}
\alias{LPG-package}
\alias{LPG}
\docType{package}
\title{
\packageTitle{LPG}
}
\description{
Fit Four-Group Probabilistic model
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{LPG}
\packageIndices{LPG}
}
\author{
\packageAuthor{LPG}
Maintainer: \packageMaintainer{LPG}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
